+++ libs/rxjs/number/src/lib/filter-out-of-range.spec.ts
@@ -1,5 +1,6 @@
 import { filterOutOfRange } from '@rxjs-ninja/rxjs-number';
 import { marbles } from 'rxjs-marbles/jest';
+import { of } from 'rxjs';
 
 describe('filterOutOfRange', () => {
   it(
@@ -20,6 +21,17 @@ describe('filterOutOfRange', () => {
       const subs = '^----------!';
       const expected = m.cold('-a-b---d-e-|', { a: -1, b: 0, d: 2, e: 3.14 });
       m.expect(input.pipe(filterOutOfRange(0, 2, true))).toBeObservable(expected);
+      m.expect(input).toHaveSubscriptions(subs);
+    }),
+  );
+
+  it(
+    'should filter values including the Observable boundary values',
+    marbles((m) => {
+      const input = m.hot('-a-b-c-d-e-|', { a: -1, b: 0, c: 1, d: 2, e: 3.14 });
+      const subs = '^----------!';
+      const expected = m.cold('-a-b---d-e-|', { a: -1, b: 0, d: 2, e: 3.14 });
+      m.expect(input.pipe(filterOutOfRange(of(0), of(2), of(true)))).toBeObservable(expected);
       m.expect(input).toHaveSubscriptions(subs);
     }),
   );