+++ libs/rxjs/number/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- More internal improvements to operators
+- All operators with inputs now accept Observable inputs
+
 ## [4.2.0] - 2021-01-17
 
 ### Changed